+++ src/test/apps/subform-test/App/logic/ModelValidator.cs
@@ -1,6 +1,3 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Altinn.App.Core.Features;
 using Altinn.App.Core.Models.Validation;
 using Altinn.App.Models.model;
 using Altinn.Platform.Storage.Interface.Models;